requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
websockets==15.0.1
python-dotenv==1.1.0
google-cloud-translate==3.20.2
google-genai==1.16.0
